requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
xgboost==1.7.6
lightgbm==4.0.0
imbalanced-learn==0.11.0
flask==2.3.3
joblib==1.3.2
plotly==5.16.1
dash==2.13.0
dash-bootstrap-components==1.4.2
python-dotenv==1.0.0
pytest==7.4.0
tqdm==4.66.1
gunicorn==21.2.0
dash-daq==0.5.0
dash-extensions==1.0.4
dash-mantine-components==0.12.1
dash-iconify==0.1.2
dash-table==5.0.0
requests==2.31.0
flask-caching==2.0.2
redis==4.6.0
psutil==5.9.5
colorama==0.4.6 